# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple project with recommendation system",
        "title": "Christmas prj",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/users/data": {
            "post": {
                "description": "post request from user for present recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "User request",
                "parameters": [
                    {
                        "description": "user's preferences",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/payload.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user's preferences has been accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid user's data",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "payload.UserRequest": {
            "type": "object",
            "required": [
                "age",
                "interests",
                "sex"
            ],
            "properties": {
                "age": {
                    "type": "integer"
                },
                "interests": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "sex": {
                    "type": "string"
                }
            }
        }
    }
}